pyproject.toml
[project]
name = "azure-cost-agent"
version = "0.1.0"
description = "Conversational Azure cost optimizer — multi-agent system built on Microsoft Agent Framework"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "agent-framework>=1.0.0rc5",
    "agent-framework-azure-ai>=1.0.0rc5",
    "azure-identity>=1.15.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-resourcegraph>=8.0.0",
    "azure-mgmt-advisor>=9.0.0",
    "azure-mgmt-monitor>=6.0.0",
    "azure-mgmt-consumption>=10.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "agent-framework-ag-ui>=1.0.0b260319",
    "azure-monitor-opentelemetry>=1.0.0",
    "chainlit>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.0",
    "mypy>=1.19.0",
    "agent-framework-devui>=1.0.0b260319",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "S", "SIM", "TCH", "RUF", "PTH", "PERF", "LOG"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.uv]
prerelease = "allow"

[tool.mypy]
python_version = "3.13"
strict = true

[[tool.mypy.overrides]]
module = [
    "azure.monitor.opentelemetry",
    "azure.ai.agentserver.agentframework",
]
ignore_missing_imports = true
